+++ modules/apps/headless-apio/person/person-apio-impl/src/main/java/com/liferay/person/apio/internal/architect/resource/PersonCollectionResource.java
@@ -14,6 +14,8 @@
 
 package com.liferay.person.apio.internal.architect.resource;
 
+import static com.liferay.portal.apio.idempotent.Idempotent.idempotent;
+
 import com.liferay.apio.architect.functional.Try;
 import com.liferay.apio.architect.pagination.PageItems;
 import com.liferay.apio.architect.pagination.Pagination;
@@ -24,23 +26,29 @@ import com.liferay.apio.architect.routes.ItemRoutes;
 import com.liferay.person.apio.architect.identifier.PersonIdentifier;
 import com.liferay.person.apio.internal.architect.form.PersonCreatorForm;
 import com.liferay.person.apio.internal.architect.form.PersonUpdaterForm;
+import com.liferay.person.apio.internal.model.UserWrapper;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.apio.architect.context.auth.MockPermissions;
-import com.liferay.portal.kernel.exception.NoSuchUserException;
+import com.liferay.portal.apio.permission.HasPermission;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Contact;
+import com.liferay.portal.kernel.model.ListType;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.UserConstants;
-import com.liferay.portal.kernel.security.auth.PrincipalException;
+import com.liferay.portal.kernel.service.ListTypeLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.service.UserService;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.LocaleUtil;
 
 import java.util.Date;
 import java.util.List;
-
-import javax.ws.rs.NotFoundException;
-import javax.ws.rs.ServerErrorException;
+import java.util.Locale;
+import java.util.function.BiFunction;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -48,7 +56,7 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * Provides the information necessary to expose <a
  * href="http://schema.org/Person">Person </a> resources through a web API. The
- * resources are mapped from the internal model {@code User}.
+ * resources are mapped from the internal model {@link UserWrapper}.
  *
  * @author Alejandro Hernández
  * @author Carlos Sierra Andrés
@@ -56,16 +64,16 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(immediate = true)
 public class PersonCollectionResource
-	implements CollectionResource<User, Long, PersonIdentifier> {
+	implements CollectionResource<UserWrapper, Long, PersonIdentifier> {
 
 	@Override
-	public CollectionRoutes<User> collectionRoutes(
-		CollectionRoutes.Builder<User> builder) {
+	public CollectionRoutes<UserWrapper> collectionRoutes(
+		CollectionRoutes.Builder<UserWrapper> builder) {
 
 		return builder.addGetter(
-			this::_getPageItems, Company.class
+			this::_getPageItems, ThemeDisplay.class
 		).addCreator(
-			this::_addUser, Company.class, MockPermissions::validPermission,
+			this::_addUser, ThemeDisplay.class, _hasPermission::forAddingUsers,
 			PersonCreatorForm::buildForm
 		).build();
 	}
@@ -76,22 +84,23 @@ public class PersonCollectionResource
 	}
 
 	@Override
-	public ItemRoutes<User, Long> itemRoutes(
-		ItemRoutes.Builder<User, Long> builder) {
+	public ItemRoutes<UserWrapper, Long> itemRoutes(
+		ItemRoutes.Builder<UserWrapper, Long> builder) {
 
 		return builder.addGetter(
-			this::_getUser
+			this::_getUserWrapper, ThemeDisplay.class
 		).addRemover(
-			this::_deleteUser, MockPermissions::validPermission
+			idempotent(_userService::deleteUser),
+			_hasPermission.forDeleting(User.class)
 		).addUpdater(
-			this::_updateUser, MockPermissions::validPermission,
-			PersonUpdaterForm::buildForm
+			this::_updateUser, ThemeDisplay.class,
+			_hasPermission.forUpdating(User.class), PersonUpdaterForm::buildForm
 		).build();
 	}
 
 	@Override
-	public Representor<User, Long> representor(
-		Representor.Builder<User, Long> builder) {
+	public Representor<UserWrapper> representor(
+		Representor.Builder<UserWrapper, Long> builder) {
 
 		return builder.types(
 			"Person"
@@ -99,10 +108,16 @@ public class PersonCollectionResource
 			User::getUserId
 		).addDate(
 			"birthDate", PersonCollectionResource::_getBirthday
+		).addLocalizedStringByLocale(
+			"honorificPrefix", _getContactField(Contact::getPrefixId)
+		).addLocalizedStringByLocale(
+			"honorificSuffix", _getContactField(Contact::getSuffixId)
 		).addString(
 			"additionalName", User::getMiddleName
 		).addString(
 			"alternateName", User::getScreenName
+		).addString(
+			"dashboardURL", UserWrapper::getDashboardURL
 		).addString(
 			"email", User::getEmailAddress
 		).addString(
@@ -112,88 +127,109 @@ public class PersonCollectionResource
 		).addString(
 			"givenName", User::getFirstName
 		).addString(
+			"image", UserWrapper::getPortraitURL
+		).addString(
 			"jobTitle", User::getJobTitle
 		).addString(
 			"name", User::getFullName
+		).addString(
+			"profileURL", UserWrapper::getProfileURL
 		).build();
 	}
 
 	private static Date _getBirthday(User user) {
-		Try<Date> dateTry = Try.fromFallible(user::getBirthday);
-
-		return dateTry.orElse(null);
+		return Try.fromFallible(
+			user::getBirthday
+		).orElse(
+			null
+		);
 	}
 
 	private static String _getGender(User user) {
-		Try<Boolean> booleanTry = Try.fromFallible(user::isMale);
-
-		return booleanTry.map(
+		return Try.fromFallible(
+			user::isMale
+		).map(
 			male -> male ? "male" : "female"
 		).orElse(
 			null
 		);
 	}
 
-	private User _addUser(
-		PersonCreatorForm personCreatorForm, Company company) {
+	private UserWrapper _addUser(
+			PersonCreatorForm personCreatorForm, ThemeDisplay themeDisplay)
+		throws PortalException {
 
-		try {
-			return _userLocalService.addUser(
-				UserConstants.USER_ID_DEFAULT, company.getCompanyId(), false,
-				personCreatorForm.getPassword1(),
-				personCreatorForm.getPassword2(),
-				personCreatorForm.hasAlternateName(),
-				personCreatorForm.getAlternateName(),
-				personCreatorForm.getEmail(), 0, StringPool.BLANK,
-				LocaleUtil.getDefault(), personCreatorForm.getGivenName(),
-				StringPool.BLANK, personCreatorForm.getFamilyName(), 0, 0,
-				personCreatorForm.isMale(),
-				personCreatorForm.getBirthdayMonth(),
-				personCreatorForm.getBirthdayDay(),
-				personCreatorForm.getBirthdayYear(),
-				personCreatorForm.getJobTitle(), null, null, null, null, false,
-				new ServiceContext());
-		}
-		catch (PortalException pe) {
-			throw new ServerErrorException(500, pe);
-		}
+		User user = _userLocalService.addUser(
+			UserConstants.USER_ID_DEFAULT, themeDisplay.getCompanyId(), false,
+			personCreatorForm.getPassword1(), personCreatorForm.getPassword2(),
+			personCreatorForm.hasAlternateName(),
+			personCreatorForm.getAlternateName(), personCreatorForm.getEmail(),
+			0, StringPool.BLANK, LocaleUtil.getDefault(),
+			personCreatorForm.getGivenName(), StringPool.BLANK,
+			personCreatorForm.getFamilyName(), 0, 0, personCreatorForm.isMale(),
+			personCreatorForm.getBirthdayMonth(),
+			personCreatorForm.getBirthdayDay(),
+			personCreatorForm.getBirthdayYear(),
+			personCreatorForm.getJobTitle(), null, null, null, null, false,
+			new ServiceContext());
+
+		return new UserWrapper(user, themeDisplay);
 	}
 
-	private void _deleteUser(Long userId) {
-		try {
-			_userLocalService.deleteUser(userId);
-		}
-		catch (PortalException pe) {
-			throw new ServerErrorException(500, pe);
-		}
+	private BiFunction<UserWrapper, Locale, String> _getContactField(
+		Function<Contact, Long> function) {
+
+		return (user, locale) -> Try.fromFallible(
+			user::getContact
+		).map(
+			function::apply
+		).map(
+			_listTypeLocalService::getListType
+		).map(
+			ListType::getName
+		).map(
+			name -> LanguageUtil.get(locale, name)
+		).orElse(
+			null
+		);
 	}
 
-	private PageItems<User> _getPageItems(
-		Pagination pagination, Company company) {
+	private PageItems<UserWrapper> _getPageItems(
+			Pagination pagination, ThemeDisplay themeDisplay)
+		throws PortalException {
 
-		List<User> users = _userLocalService.getCompanyUsers(
-			company.getCompanyId(), pagination.getStartPosition(),
-			pagination.getEndPosition());
-		int count = _userLocalService.getCompanyUsersCount(
-			company.getCompanyId());
+		List<UserWrapper> userWrappers = Stream.of(
+			_userService.getCompanyUsers(
+				themeDisplay.getCompanyId(), pagination.getStartPosition(),
+				pagination.getEndPosition())
+		).flatMap(
+			List::stream
+		).map(
+			user -> new UserWrapper(user, themeDisplay)
+		).collect(
+			Collectors.toList()
+		);
 
-		return new PageItems<>(users, count);
+		int count = _userService.getCompanyUsersCount(
+			themeDisplay.getCompanyId());
+
+		return new PageItems<>(userWrappers, count);
 	}
 
-	private User _getUser(Long userId) {
-		try {
-			return _userLocalService.getUserById(userId);
-		}
-		catch (NoSuchUserException | PrincipalException e) {
-			throw new NotFoundException("Unable to get user " + userId, e);
-		}
-		catch (PortalException pe) {
-			throw new ServerErrorException(500, pe);
-		}
+	private UserWrapper _getUserWrapper(long userId, ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		User user = _userService.getUserById(userId);
+
+		return new UserWrapper(user, themeDisplay);
 	}
 
-	private User _updateUser(Long userId, PersonUpdaterForm personUpdaterForm) {
-		User user = _getUser(userId);
+	private UserWrapper _updateUser(
+			long userId, PersonUpdaterForm personUpdaterForm,
+			ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		User user = _userService.getUserById(userId);
 
 		user.setPassword(personUpdaterForm.getPassword());
 		user.setScreenName(personUpdaterForm.getAlternateName());
@@ -202,13 +238,21 @@ public class PersonCollectionResource
 		user.setLastName(personUpdaterForm.getFamilyName());
 		user.setJobTitle(personUpdaterForm.getJobTitle());
 
-		Try<User> userTry = Try.fromFallible(
-			() -> _userLocalService.updateUser(user));
+		user = _userLocalService.updateUser(user);
 
-		return userTry.getUnchecked();
+		return new UserWrapper(user, themeDisplay);
 	}
 
 	@Reference
+	private HasPermission _hasPermission;
+
+	@Reference
+	private ListTypeLocalService _listTypeLocalService;
+
+	@Reference
 	private UserLocalService _userLocalService;
+
+	@Reference
+	private UserService _userService;
 
 }