+++ modules/apps/static/portal-osgi-web/portal-osgi-web-portlet-tracker/src/main/java/com/liferay/portal/osgi/web/portlet/tracker/internal/PortletTracker.java
@@ -273,6 +273,28 @@ public class PortletTracker
 			BundlePortletApp bundlePortletApp = createBundlePortletApp(
 				bundle, bundleClassLoader, serviceRegistrations);
 
+			bundlePortletApp.setDefaultNamespace(
+				(String)serviceReference.getProperty(
+					"javax.portlet.default-namespace"));
+
+			String jxPortletVersion = (String)serviceReference.getProperty(
+				"javax.portlet.version");
+
+			if (jxPortletVersion != null) {
+				String[] jxPortletVersionParts = StringUtil.split(
+					jxPortletVersion, CharPool.PERIOD);
+
+				if (jxPortletVersionParts.length > 0) {
+					bundlePortletApp.setSpecMajorVersion(
+						GetterUtil.getInteger(jxPortletVersionParts[0], 2));
+
+					if (jxPortletVersionParts.length > 1) {
+						bundlePortletApp.setSpecMinorVersion(
+							GetterUtil.getInteger(jxPortletVersionParts[1]));
+					}
+				}
+			}
+
 			com.liferay.portal.kernel.model.Portlet portletModel =
 				buildPortletModel(bundlePortletApp, portletId);
 
@@ -554,10 +576,6 @@ public class PortletTracker
 			GetterUtil.getBoolean(
 				get(serviceReference, "private-session-attributes"),
 				portletModel.isPrivateSessionAttributes()));
-		portletModel.setRemoteable(
-			GetterUtil.getBoolean(
-				get(serviceReference, "remoteable"),
-				portletModel.isRemoteable()));
 		portletModel.setRenderTimeout(
 			GetterUtil.getInteger(
 				get(serviceReference, "render-timeout"),